backport from php8.2

+++ source/opencv2/core/opencv_mat.cc
@@ -194,7 +194,7 @@ PHP_METHOD(opencv_mat, createWithDims)
         RETURN_NULL();
     }
 
-    int sizes_arr[dims];
+    int sizes_arr[(const long)dims];
     HashTable *sizes_ht = Z_ARRVAL_P(sizes_zval);
 
     if (zend_hash_num_elements(sizes_ht) < dims)
